requirements.txt
streamlit>=1.28.0
anthropic>=0.18.0
python-dotenv>=1.0.0
pandas>=2.0.0
plotly>=5.18.0

# LangChain ecosystem
langchain>=0.1.0
langchain-anthropic>=0.1.0
langgraph>=0.0.40
pydantic>=2.0.0

# Audio recording and transcription
audio-recorder-streamlit>=0.0.8
openai>=1.0.0
